refacto: split import data from DS to PG to 2 files
